requirements.txt
python>=3.9
Pillow==10.1.0
